Move command line parsing to parse_args.rs
